Separated Enu, LatLon and Ecef in different files
